Implement useContext in TimelineComponent
when clicking on a timeline content the content should stay zoomed in.
to achieve this we need to update styles on the basis on state. Introduced
useContext to manage and use the state wherever required hence reducing the
tedious task to pass state data by prop drilling through multiple components.